+++ Classes/Model/EmptyFPDI.php
@@ -36,6 +36,14 @@ use FPDI;
  */
 class EmptyFPDI extends FPDI {
 	/**
+	 * Indicating whether the current page is using a template.
+	 * This is needed to automatically add the template on an automatic page break.
+	 *
+	 * @var bool
+	 */
+	protected $importTemplateOnThisPage = false;
+
+	/**
 	 * @return void
 	 */
 	public function Header() {
@@ -47,5 +55,28 @@ class EmptyFPDI extends FPDI {
 	 */
 	public function Footer() {
 
+	}
+
+	/**
+	 * Fixes importPage not working with autoPageBreak=1, see https://github.com/bithost-gmbh/pdfviewhelpers/issues/41
+	 *
+	 * @inheritdoc
+	 */
+	public function AddPage($orientation = '', $format = '', $rotationOrKeepmargins = false, $tocpage = false)
+	{
+		parent::AddPage($orientation, $format, $rotationOrKeepmargins, $tocpage);
+
+		if ($this->importTemplateOnThisPage && $this->tpl !== 0) {
+			$this->useTemplate($this->tpl);
+		}
+	}
+
+	/**
+	 * @param bool $importTemplateOnThisPage
+	 *
+	 * @return void
+	 */
+	public function setImportTemplateOnThisPage($importTemplateOnThisPage) {
+		$this->importTemplateOnThisPage = $importTemplateOnThisPage;
 	}
 }